fix: error on base64pad
I'm not wise enough to know if this change is a good idea, but base64pad strings as bytes values change the encoded form and are forbidden per the spec.

Notably it will make some exiting IPNI advertisements unencodable. But in principle, the encoder should probably warn the user when they try an encode an illegal bytes string.

> The Base64 encoding is the one described in [RFC 4648, section 4](https://tools.ietf.org/html/rfc4648#section-4) **without padding.**
> – https://ipld.io/specs/codecs/dag-json/spec/#bytes

fixes: https://github.com/ipld/js-dag-json/issues/106

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>